+++ payments-core/src/test/java/com/stripe/android/networking/StripeApiRepositoryTest.kt
@@ -1189,7 +1189,9 @@ internal class StripeApiRepositoryTest {
                 TokenFixtures.CARD_TOKEN_JSON.toString(),
                 emptyMap()
             )
-            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(stripeResponse)
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(
+                stripeResponse
+            )
 
             val productUsage = "TestProductUsage"
             create(setOf(productUsage)).createToken(
@@ -1213,7 +1215,9 @@ internal class StripeApiRepositoryTest {
                 TokenFixtures.CARD_TOKEN_JSON.toString(),
                 emptyMap()
             )
-            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(stripeResponse)
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(
+                stripeResponse
+            )
 
             val productUsage = "TestProductUsage"
             create(setOf(productUsage)).createToken(
@@ -1238,7 +1242,9 @@ internal class StripeApiRepositoryTest {
                 TokenFixtures.CARD_TOKEN_JSON.toString(),
                 emptyMap()
             )
-            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(stripeResponse)
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(
+                stripeResponse
+            )
             create().createToken(
                 CardParamsFixtures.WITH_ATTRIBUTION,
                 DEFAULT_OPTIONS
@@ -1395,7 +1401,9 @@ internal class StripeApiRepositoryTest {
                 PaymentMethodFixtures.SEPA_DEBIT_JSON.toString(),
                 emptyMap()
             )
-            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(stripeResponse)
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>())).thenReturn(
+                stripeResponse
+            )
             create().createPaymentMethod(
                 PaymentMethodCreateParams.create(
                     PaymentMethodCreateParams.SepaDebit("my_iban")
@@ -1590,6 +1598,97 @@ internal class StripeApiRepositoryTest {
             verify(stripeNetworkClient).executeRequest(apiRequestArgumentCaptor.capture())
             val params = requireNotNull(apiRequestArgumentCaptor.firstValue.params)
             assertEquals(params["email_address"], email)
+        }
+
+    @Test
+    fun `consumerSignUp() sends all parameters`() =
+        runTest {
+            val stripeResponse = StripeResponse(
+                200,
+                ConsumerFixtures.CONSUMER_VERIFIED_JSON.toString(),
+                emptyMap()
+            )
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>()))
+                .thenReturn(stripeResponse)
+
+            val email = "email@example.com"
+            val phoneNumber = "phone number"
+            val country = "US"
+            val cookies = "cookie1,cookie2"
+            create().consumerSignUp(
+                email,
+                phoneNumber,
+                country,
+                cookies,
+                DEFAULT_OPTIONS
+            )
+
+            verify(stripeNetworkClient).executeRequest(apiRequestArgumentCaptor.capture())
+            val params = requireNotNull(apiRequestArgumentCaptor.firstValue.params)
+            assertEquals(params["email_address"], email)
+            assertEquals(params["phone_number"], phoneNumber)
+            assertEquals(params["country"], country)
+            assertEquals(params["cookies"], cookies)
+        }
+
+    @Test
+    fun `startConsumerVerification() sends all parameters`() =
+        runTest {
+            val stripeResponse = StripeResponse(
+                200,
+                ConsumerFixtures.CONSUMER_VERIFICATION_STARTED_JSON.toString(),
+                emptyMap()
+            )
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>()))
+                .thenReturn(stripeResponse)
+
+            val clientSecret = "secret"
+            val locale = Locale.US
+            val cookies = "cookie1,cookie2"
+            create().startConsumerVerification(
+                clientSecret,
+                locale,
+                cookies,
+                DEFAULT_OPTIONS
+            )
+
+            verify(stripeNetworkClient).executeRequest(apiRequestArgumentCaptor.capture())
+            val params = requireNotNull(apiRequestArgumentCaptor.firstValue.params)
+            val credentials = params["credentials"] as Map<*, *>
+            assertEquals(credentials["consumer_session_client_secret"], clientSecret)
+            assertEquals(params["type"], "SMS")
+            assertEquals(params["locale"], locale.toLanguageTag())
+            assertEquals(params["cookies"], cookies)
+        }
+
+    @Test
+    fun `confirmConsumerVerification() sends all parameters`() =
+        runTest {
+            val stripeResponse = StripeResponse(
+                200,
+                ConsumerFixtures.CONSUMER_VERIFIED_JSON.toString(),
+                emptyMap()
+            )
+            whenever(stripeNetworkClient.executeRequest(any<ApiRequest>()))
+                .thenReturn(stripeResponse)
+
+            val clientSecret = "secret"
+            val verificationCode = "1234"
+            val cookies = "cookie1,cookie2"
+            create().confirmConsumerVerification(
+                clientSecret,
+                verificationCode,
+                cookies,
+                DEFAULT_OPTIONS
+            )
+
+            verify(stripeNetworkClient).executeRequest(apiRequestArgumentCaptor.capture())
+            val params = requireNotNull(apiRequestArgumentCaptor.firstValue.params)
+            val credentials = params["credentials"] as Map<*, *>
+            assertEquals(credentials["consumer_session_client_secret"], clientSecret)
+            assertEquals(params["type"], "SMS")
+            assertEquals(params["code"], verificationCode)
+            assertEquals(params["cookies"], cookies)
         }
 
     private fun verifyFraudDetectionDataAndAnalyticsRequests(