+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 ## X.X.X - 2022-XX-XX
+[FIXED] [4606](https://github.com/stripe/stripe-android/pull/4606) Keep status bar color in PaymentLauncher.
+### PaymentSheet
+### Identity
+### Card scanning
+[ADDED] [4592](https://github.com/stripe/stripe-android/pull/4592) Add support for launching card scan from fragments.
+
+## 19.2.0 - 2022-02-14
+This release includes several bug fixes and upgrades Kotlin to 1.6.
 * [CHANGED] [4546](https://github.com/stripe/stripe-android/pull/4546) Update to kotlin 1.6
 * [FIXED] [4560](https://github.com/stripe/stripe-android/pull/4560) Fix `cardValidCallback` being added multiple times in `CardInputWidget`.
 * [FIXED] [4574](https://github.com/stripe/stripe-android/pull/4574) Take `postalCode` into account in `CardMultilineWidget` validation.
@@ -7,7 +15,6 @@
 ### PaymentSheet
 * [FIXED] [4466](https://github.com/stripe/stripe-android/pull/4466) Fix issues when activities are lost on low resource phones.
 * [FIXED] [4557](https://github.com/stripe/stripe-android/pull/4557) Add missing app info to some Stripe API requests
-### Identity
 ### Card scanning
 * [FIXED] [4548](https://github.com/stripe/stripe-android/pull/4548) Potential work leak when canceling a card scan in StripeCardScan
 * [ADDED] [4562](https://github.com/stripe/stripe-android/pull/4562) Add an example page for cardscan